Renames packages to align with other fedora ecosystem projects.
