+++ pom.xml
@@ -2,8 +2,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.ibr.fedora</groupId>
-  <artifactId>testSuite</artifactId>
+  <groupId>org.fcrepo.spec</groupId>
+  <artifactId>testsuite</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
 
@@ -33,7 +33,7 @@
     <project.copyrightYear>2017</project.copyrightYear>
     <license.plugin.version>2.11</license.plugin.version>
     <javax.xml.version>1.4.5</javax.xml.version>
-    <app.main.class>com.ibr.fedora.App</app.main.class>
+    <app.main.class>org.fcrepo.spec.testsuite.App</app.main.class>
   </properties>
 
   <build>
