+++ src/main/java/org/fcrepo/spec/testsuite/test/HttpDelete.java
@@ -18,13 +18,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ibr.fedora.testsuite;
+package org.fcrepo.spec.testsuite.test;
 
 import java.io.FileNotFoundException;
 import java.io.PrintStream;
 
-import com.ibr.fedora.TestSuiteGlobals;
-import com.ibr.fedora.TestsLabels;
+import org.fcrepo.spec.testsuite.TestSuiteGlobals;
+import org.fcrepo.spec.testsuite.TestsLabels;
 import io.restassured.RestAssured;
 import io.restassured.config.LogConfig;
 import io.restassured.http.Header;
